+++ src/components/common/coloborations.ts
@@ -1,0 +1,33 @@
+interface coloboration {
+  name: string,
+  github: string,
+  create: string[]
+}
+
+export const coloborations: coloboration[] = [
+  {
+    name: "Aleksei Zhuchkov",
+    github: "https://github.com/kelzerock",
+    create: [
+      "something",
+      "something1",
+    ]
+  },
+  {
+    name: "Ihar Pavetka",
+    github: "https://github.com/MuJlblii",
+    create: [
+      "something",
+      "something1",
+    ]
+  },
+  {
+    name: "Aleksei Artsymenya",
+    github: "https://github.com/Alex-Artsymenya",
+    create: [
+      "something",
+      "something1",
+    ]
+  },
+
+]
