+++ src/index.html
@@ -12,8 +12,10 @@
       rel="stylesheet"/>
   </head>
   <body>
-    <header id="header_container" class="header"></header>
-    <main id="page_container"></main>
-    <footer id="footer_container" class="footer"></footer>
+    <div class="main-wrapper">
+      <header id="header_container" class="header"></header>
+      <main id="page_container"></main>
+      <footer id="footer_container" class="footer"></footer>
+    </div>
   </body>
 </html>
